requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-ckeditor==6.7.0
django-colorfield==0.9.0
django-js-asset==2.1.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.2.0
Pillow==10.0.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
